Arrumando o HealthUsu para salvar tambem o TDEE

+++ src/main/java/br/com/nutriscare/nutrisCareRulesApi/dto/UserDTO.java
@@ -33,6 +33,6 @@ public class UserDTO {
 
     private Double height;
     private Double weight;
-    private Double levelGoals;
-    private Double levelPhysicalActivityLevel;
+    private Integer levelGoals;
+    private Integer levelPhysicalActivityLevel;
 }
